# Optimize batch inference for text generation