--- a/tests/integration/llm/client.py
+++ b/tests/integration/llm/client.py
@@ -247,7 +247,9 @@ def test_handler(model, model_spec):
             if spec.get("stream_output", False):
                 logging.info(f"res: {res.content}")
                 result = res.content.decode().split("\n")[:-1]
-                assert len(result) <= seq_length, "generated more takens than max_new_tokens"
+                assert len(
+                    result
+                ) <= seq_length, "generated more takens than max_new_tokens"
             else:
                 res = res.json()
                 logging.info(f"res {res}")